Backend: Refactor router to its own file.
