+++ robot-server/tests/runs/router/test_base_router.py
@@ -25,7 +25,6 @@ from robot_server.data_files.data_files_store import (
 
 from robot_server.data_files.models import DataFileSource
 from robot_server.errors.error_responses import ApiError
-from robot_server.runs.error_recovery_models import ErrorRecoveryPolicy
 from robot_server.service.json_api import (
     RequestModel,
     SimpleBody,
@@ -67,7 +66,6 @@ from robot_server.runs.router.base_router import (
     get_runs,
     remove_run,
     update_run,
-    put_error_recovery_policy,
     get_run_commands_error,
     get_current_state,
     CurrentStateLinks,
@@ -699,44 +697,6 @@ async def test_update_to_current_missing(
 
     assert exc_info.value.status_code == 404
     assert exc_info.value.content["errors"][0]["id"] == "RunNotFound"
-
-
-async def test_create_policies(
-    decoy: Decoy, mock_run_data_manager: RunDataManager
-) -> None:
-    """It should call RunDataManager create run policies."""
-    policies = decoy.mock(cls=ErrorRecoveryPolicy)
-    await put_error_recovery_policy(
-        runId="rud-id",
-        request_body=RequestModel(data=policies),
-        run_data_manager=mock_run_data_manager,
-    )
-    decoy.verify(
-        mock_run_data_manager.set_error_recovery_rules(
-            run_id="rud-id", rules=policies.policyRules
-        )
-    )
-
-
-async def test_create_policies_raises_not_active_run(
-    decoy: Decoy, mock_run_data_manager: RunDataManager
-) -> None:
-    """It should raise that the run is not current."""
-    policies = decoy.mock(cls=ErrorRecoveryPolicy)
-    decoy.when(
-        mock_run_data_manager.set_error_recovery_rules(
-            run_id="rud-id", rules=policies.policyRules
-        )
-    ).then_raise(RunNotCurrentError())
-    with pytest.raises(ApiError) as exc_info:
-        await put_error_recovery_policy(
-            runId="rud-id",
-            request_body=RequestModel(data=policies),
-            run_data_manager=mock_run_data_manager,
-        )
-
-    assert exc_info.value.status_code == 409
-    assert exc_info.value.content["errors"][0]["id"] == "RunStopped"
 
 
 async def test_get_run_commands_errors(
